requirements-pi.txt
# Core dependencies (lighter versions for Pi)
PyQt5>=5.15.0
PyOpenGL>=3.1.6
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
sounddevice>=0.4.6

# Speech Recognition (quantized models)
faster-whisper>=1.0.0

# Text-to-Speech - Voice Cloning (RVC + Base TTS)
# RVC Voice Conversion
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
faiss-cpu>=1.7.0

# Base TTS Engine
edge-tts>=6.1.0

# Audio processing
pyaudio>=0.2.13
scipy>=1.11.0
pygame>=2.5.0  # For playing generated audio files

# Database
aiosqlite>=0.19.0

# Utilities
requests>=2.31.0

